标注文件中坐标采用浮点数计算，避免小标注框发生位移和精度损失

+++ utils/ToXML5D0.py
@@ -52,21 +52,13 @@ class CreatSaveXml():
 
             bndbox = ET.SubElement(object, "bndbox")
             if using == "dcm":
-                # ET.SubElement(bndbox, "xmin").text = str(int(pixmapItem.mapFromScene(
-                #     items[len(items) - 2 - index].mapToScene(items[len(items) - 2 - index].rect().topLeft())).x()))
-                # ET.SubElement(bndbox, "ymin").text = str(int(pixmapItem.mapFromScene(
-                #     items[len(items) - 2 - index].mapToScene(items[len(items) - 2 - index].rect().topLeft())).y()))
-                # ET.SubElement(bndbox, "xmax").text = str(int(pixmapItem.mapFromScene(
-                #     items[len(items) - 2 - index].mapToScene(items[len(items) - 2 - index].rect().bottomRight())).x()))
-                # ET.SubElement(bndbox, "ymax").text = str(int(pixmapItem.mapFromScene(
-                #     items[len(items) - 2 - index].mapToScene(items[len(items) - 2 - index].rect().bottomRight())).y()))
-                ET.SubElement(bndbox, "xmin").text = str(int(pixmapItem.mapFromScene(
+                ET.SubElement(bndbox, "xmin").text = str(float(pixmapItem.mapFromScene(
                     item.mapToScene(item.rect().topLeft())).x()))
-                ET.SubElement(bndbox, "ymin").text = str(int(pixmapItem.mapFromScene(
+                ET.SubElement(bndbox, "ymin").text = str(float(pixmapItem.mapFromScene(
                     item.mapToScene(item.rect().topLeft())).y()))
-                ET.SubElement(bndbox, "xmax").text = str(int(pixmapItem.mapFromScene(
+                ET.SubElement(bndbox, "xmax").text = str(float(pixmapItem.mapFromScene(
                     item.mapToScene(item.rect().bottomRight())).x()))
-                ET.SubElement(bndbox, "ymax").text = str(int(pixmapItem.mapFromScene(
+                ET.SubElement(bndbox, "ymax").text = str(float(pixmapItem.mapFromScene(
                     item.mapToScene(item.rect().bottomRight())).y()))
             else:
                 print(pixmapItem)
